Stop parsing replay at requested cutoff time
Previously, the whole replay was parsed and then each tracker had to trim its
data records to the requested cutoff time (the --until flag). Now we just stop
parsing once we get past that time.

Also, made a couple of tweaks to better align timestamps on x-axes for all trackers.